@@ -14,9 +14,10 @@
   "devDependencies": {
     "@babel/core": "^7.13.14",
     "@babel/preset-env": "^7.13.12",
+    "@vue/compiler-sfc": "^3.0.11",
     "autoprefixer": "^10.2.5",
     "babel-loader": "^8.2.2",
-    "babel-plugin-component": "^1.1.1",
+    "babel-plugin-import": "^1.13.3",
     "clean-webpack-plugin": "^3.0.0",
     "compression-webpack-plugin": "^7.1.2",
     "css-loader": "^5.2.0",
@@ -26,8 +27,7 @@
     "postcss-preset-env": "^6.7.0",
     "style-loader": "^2.0.0",
     "tailwindcss": "^2.1.1",
-    "vue-loader": "^15.9.6",
-    "vue-template-compiler": "^2.6.12",
+    "vue-loader": "^16.2.0",
     "webpack": "^5.30.0",
     "webpack-cli": "^4.6.0",
     "webpack-dev-server": "^4.0.0-beta.2",
@@ -35,8 +35,8 @@
   },
   "dependencies": {
     "core-js": "^3.10.0",
-    "element-ui": "^2.15.1",
-    "vue": "^2.6.12",
-    "vue-router": "^3.5.1"
+    "element-plus": "^1.0.2-beta.36",
+    "vue": "^3.0.11",
+    "vue-router": "^4.0.6"
   }
 }
