@@ -1,6 +1,7 @@
 const path = require('path')
 const HtmlWebpackPlugin = require('html-webpack-plugin')
-const VueLoaderPlugin = require('vue-loader/lib/plugin')
+const { VueLoaderPlugin } = require('vue-loader')
+const { DefinePlugin } = require('webpack')
 module.exports = {
   resolve: {
     alias: {
@@ -31,6 +32,11 @@ module.exports = {
       template: './src/index.html',
       favicon: './assets/favicon.ico',
       hash: true, // HTML引入资源时，在文件名后拼接hash
+    }),
+    // vlevin @ 2021-04-07 可以不配，但文档强烈建议配置，参见https://github.com/vuejs/vue-next/tree/master/packages/vue#bundler-build-feature-flags
+    new DefinePlugin({
+      __VUE_OPTIONS_API__: true,
+      __VUE_PROD_DEVTOOLS__: false,
     }),
     // vlevin @ 2021-02-17 待修复，不知为何在Vue下不可，在React下可
     // new webpack.ProvidePlugin({
